migrationed type declarations to a types.go file
